requirements.txt
# File Logic Summary: Python dependency manifest required by backend and ML components.
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
email-validator==2.1.0

# Audio Processing
librosa==0.10.0
soundfile==0.12.1
whisper-openai==20240314
ffmpeg-python==0.2.1

# ML Models
torch==2.1.1
transformers==4.35.2
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.4

# PDF Generation
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
corsheaders==4.3.1
requests==2.31.0

# Testing/Development
pytest==7.4.3

